requirements.txt
rich==13.7.0
typer==0.9.0
prompt-toolkit==3.0.43
colorama==0.4.6
shellingham==1.5.4
openai>=1.79.0
duckduckgo-search>=8.0.0
crewai>=0.28.0
crewai[tools]>=0.28.0
langchain-openai
python-dotenv
setuptools
beautifulsoup4
requests